+++ trunk/Server/Stump.Server.WorldServer/Game/Fights/Triggers/Trap.cs
@@ -13,14 +13,14 @@ namespace Stump.Server.WorldServer.Game.Fights.Triggers
 {
     public class Trap : MarkTrigger
     {
-        public Trap(short id, FightActor caster, Spell castedSpell, EffectDice originEffect, Spell glyphSpell, Cell centerCell, sbyte size)
+        public Trap(short id, FightActor caster, Spell castedSpell, EffectDice originEffect, Spell glyphSpell, Cell centerCell, byte size)
             : base(id, caster, castedSpell, originEffect, new MarkShape(caster.Fight, centerCell, GameActionMarkCellsTypeEnum.CELLS_CIRCLE, size, Color.Brown))
         {
             TrapSpell = glyphSpell;
             VisibleState = GameActionFightInvisibilityStateEnum.INVISIBLE;
         }
 
-        public Trap(short id, FightActor caster, Spell spell, EffectDice originEffect, Spell trapSpell, Cell centerCell, GameActionMarkCellsTypeEnum shape, sbyte size)
+        public Trap(short id, FightActor caster, Spell spell, EffectDice originEffect, Spell trapSpell, Cell centerCell, GameActionMarkCellsTypeEnum shape, byte size)
             : base(id, caster, spell, originEffect, new MarkShape(caster.Fight, centerCell, shape, size, Color.Brown))
         {
             TrapSpell = trapSpell;
@@ -70,7 +70,7 @@ namespace Stump.Server.WorldServer.Game.Fights.Triggers
                 foreach (var effect in TrapSpell.CurrentSpellLevel.Effects)
                 {
                     var handler = EffectManager.Instance.GetSpellEffectHandler(effect, Caster, TrapSpell, shape.Cell, false);
-                    handler.EffectZone = new Zone(shape.Shape == GameActionMarkCellsTypeEnum.CELLS_CROSS ? SpellShapeEnum.Q : effect.ZoneShape, (uint)shape.Size);
+                    handler.EffectZone = new Zone(shape.Shape == GameActionMarkCellsTypeEnum.CELLS_CROSS ? SpellShapeEnum.Q : effect.ZoneShape, shape.Size);
 
                     handler.Apply();
                 }
